app/requirements.txt
absl-py==0.9.0
google-cloud-bigquery==1.26.0
google-cloud-storage==1.29.0
